requirements.txt
# Requirements for Utility Token Generation Project
# Install with: pip install -r requirements.txt

# Core dependencies
pycryptodome>=3.15.0  # For cryptographic operations (DES)

# Data processing
pandas>=1.3.0
numpy>=1.20.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Excel support
openpyxl>=3.0.0  # For Excel file generation

# GUI
tkinter>=8.6  # Usually comes with Python standard library
